requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai>=0.8.3
python-dotenv
